Rename: 웹 애플리케이션 실행 클래스, artifact 이름 수정
library 도메인을 담당할 library 패키지와 구분되게 하기 위해 디렉토리 이름변경
