+++ src/main/java/ksp7331/practice/libraryAPI/LibraryAPIApplication.java
@@ -1,13 +1,13 @@
-package ksp7331.practice.library;
+package ksp7331.practice.libraryAPI;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class LibraryApplication {
+public class LibraryAPIApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(LibraryApplication.class, args);
+		SpringApplication.run(LibraryAPIApplication.class, args);
 	}
 
 }
