+++ src/test/java/ksp7331/practice/libraryAPI/LibraryAPIApplicationTests.java
@@ -1,10 +1,10 @@
-package ksp7331.practice.library;
+package ksp7331.practice.libraryAPI;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class LibraryApplicationTests {
+class LibraryAPIApplicationTests {
 
 	@Test
 	void contextLoads() {
